+++ include/data-file-library/impl/api_impl_.h
@@ -1,0 +1,42 @@
+/* 
+    Data File Library: A set of C functions for handling numeric data files
+
+    File: include/data-file-library/impl/api_impl_.h
+    Version: include/data-file-library/version.h
+    Author: Jhonas Olivati de Sarro
+    Language standards: C99 and C++98 guard
+    License: include/data-file-library/license.txt
+
+    Description:
+        Define the API with macros for C++, and for compilation, and for
+        header-only or compiled library usage.
+*/
+
+#ifndef DATA_FILE_LIBRARY_API_IMPL_H
+#define DATA_FILE_LIBRARY_API_IMPL_H
+
+#ifdef DATA_FILE_LIBRARY_EXPORTS_IMPL_
+    #if defined(_WIN32) || defined(_WIN64)
+        #define DATA_FILE_LIBRARY_VISIBILITY_IMPL_ __declspec(dllexport)
+    #else
+        #define DATA_FILE_LIBRARY_VISIBILITY_IMPL_ __attribute__((visibility("default")))
+    #endif
+#elif defined(DATA_FILE_LIBRARY_IMPORTS)
+    #if defined(_WIN32) || defined(_WIN64)
+        #define DATA_FILE_LIBRARY_VISIBILITY_IMPL_ __declspec(dllimport)
+    #else
+        #define DATA_FILE_LIBRARY_VISIBILITY_IMPL_
+    #endif
+#else
+    #define DATA_FILE_LIBRARY_VISIBILITY_IMPL_ static inline
+#endif
+
+#ifdef __cplusplus
+    #define DATA_FILE_LIBRARY_CPP_IMPL_ extern "C"
+#else
+    #define DATA_FILE_LIBRARY_CPP_IMPL_
+#endif
+
+#define DATA_FILE_LIBRARY_API_IMPL_ DATA_FILE_LIBRARY_VISIBILITY_IMPL_ DATA_FILE_LIBRARY_CPP_IMPL_
+
+#endif /* DATA_FILE_LIBRARY_API_IMPL_H */